A program that finds and prints the sum of the even valued Fibonacci numbers below 4000000

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,21 +8,20 @@
 int main(void)
 {
 	unsigned long a = 0, b = 1, sum;
-	unsigned long total;
+	float total;
 
 	while (1)
 	{
 		sum = a + b;
 		if (sum > 4000000)
 			break;
+
 		if ((sum % 2) == 0)
 			total = total + sum;
 		
 		a = b;
 		b = sum;
 	}
-	printf("%ld\n", total);
+	printf("%.0f\n", total);
 	return (0);
 }
-
-
